Drastic undoing to minimize the diff. Some of this is likely to be restored
