--- samples/Samples.HttpClientDriver.LoaderOpt/Properties/launchSettings.json
@@ -1,21 +1,0 @@
-﻿{
-  "profiles": {
-    "Samples.HttpClientDriver.LoaderOpt": {
-      "commandName": "Project",
-      "commandLineArgs": "",
-      "environmentVariables": {
-        "COR_ENABLE_PROFILING": "1",
-        "COR_PROFILER": "{846F5F1C-F9AE-4B07-969E-05C26BC060D8}",
-        "COR_PROFILER_PATH": "$(ProjectDir)$(OutputPath)profiler-lib\\Datadog.Trace.ClrProfiler.Native.dll",
-
-        "CORECLR_ENABLE_PROFILING": "1",
-        "CORECLR_PROFILER": "{846F5F1C-F9AE-4B07-969E-05C26BC060D8}",
-        "CORECLR_PROFILER_PATH": "$(ProjectDir)$(OutputPath)profiler-lib\\Datadog.Trace.ClrProfiler.Native.dll",
-
-        "DD_DOTNET_TRACER_HOME": "$(ProjectDir)$(OutputPath)profiler-lib",
-        "DD_INTEGRATIONS": "$(ProjectDir)$(OutputPath)profiler-lib\\integrations.json"
-      },
-      "nativeDebugging": true
-    }
-  }
-}
